+++ Screens/log.html
@@ -60,8 +60,11 @@
             </div>
         </div>
     </main>
+    <script src="../Scripts/Contracts/Project.js"></script>
     <script src="../Scripts/Contracts/State.js"></script>
-    <script src="../Scripts/controllers.js"></script>
+    <script src="../Scripts/Contracts/Status.js"></script>
+    <script src="../Scripts/Contracts/Task.js"></script>
+    <script src="./renderer.js"></script>
     <script src="../Scripts/pageListeners.js"></script>
 </body>
 </html>
