+++ Scripts/Contracts/Project.js
@@ -1,8 +1,7 @@
 const Project = class {
-    constructor (id, name){
+    constructor (id, name, tasks){
         this.id = id
         this.name = name
+        this.tasks = tasks
     }
 }
-
-module.exports = { Project }
