+++ Scripts/Contracts/Status.js
@@ -6,5 +6,3 @@ const Status = class {
     static WAITING = 5
     static WONT_DO = 6
 }
-
-module.exports = { Status }
