+++ package.json
@@ -2,7 +2,7 @@
   "name": "protracker",
   "version": "1.0.0",
   "description": "app to track tasks",
-  "main": "./Scripts/mainElectron.js",
+  "main": "./Scripts/eMain.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "electron .",
